combine Versions.svelte and `OfficialVersions.svelte`, refine these files again later
