+++ src/routes/settings/versions/VersionList.svelte
@@ -6,7 +6,6 @@
   import IconDownload from "~icons/mdi/download";
   import IconDeleteForever from "~icons/mdi/delete-forever";
   import {
-    Alert,
     Button,
     Radio,
     Spinner,
@@ -25,7 +24,6 @@
   let {
     description,
     releaseList,
-    loaded,
     onVersionChange,
     onRemoveVersion,
     onRefreshVersions,
@@ -33,7 +31,6 @@
   }: {
     description: string;
     releaseList: ReleaseInfo[];
-    loaded: boolean;
     onVersionChange: (version: string) => Promise<void>;
     onRemoveVersion: (version: string) => Promise<void>;
     onRefreshVersions: () => Promise<void>;
@@ -56,145 +53,123 @@
   };
 </script>
 
-{#if !loaded}
-  <div class="flex flex-col justify-center items-center">
-    <Spinner color="yellow" size="12" />
+<div class="flex items-center mb-2">
+  <div class="grow">
+    <p class="text-sm text-gray-300">
+      {description}
+    </p>
   </div>
-{:else}
-  <div class="flex items-center mb-2">
-    <div class="grow">
-      <p class="text-sm text-gray-300">
-        {description}
-      </p>
-    </div>
-    <div class="flex">
-      <Button
-        class="p-2! mr-2 rounded-md bg-orange-500 hover:bg-orange-600 text-slate-900"
-        onclick={() => onRefreshVersions()}
-      >
-        <IconRefresh
-          aria-label={$_("settings_versions_icon_refresh_altText")}
-        />
-      </Button>
-      <Button
-        class="p-2! rounded-md bg-orange-500 hover:bg-orange-600 text-slate-900"
-        onclick={async () =>
-          openPath((await getInstallationDirectory()) + "/versions/official/")}
-      >
-        <IconFolderOpen
-          aria-label={$_("settings_versions_icon_openFolder_altText")}
-        />
-      </Button>
-    </div>
+  <div class="flex">
+    <Button
+      class="p-2! mr-2 rounded-md bg-orange-500 hover:bg-orange-600 text-slate-900"
+      onclick={() => onRefreshVersions()}
+    >
+      <IconRefresh aria-label={$_("settings_versions_icon_refresh_altText")} />
+    </Button>
+    <Button
+      class="p-2! rounded-md bg-orange-500 hover:bg-orange-600 text-slate-900"
+      onclick={async () =>
+        openPath((await getInstallationDirectory()) + "/versions/official/")}
+    >
+      <IconFolderOpen
+        aria-label={$_("settings_versions_icon_openFolder_altText")}
+      />
+    </Button>
   </div>
+</div>
 
-  {#if releaseList.length === 0}
-    <Alert class="bg-slate-900 grow text-red-400">
-      {$_("settings_versions_noReleasesFound")}
-    </Alert>
-  {:else}
-    <Table>
-      <TableHead class="bg-slate-400">
-        <TableHeadCell></TableHeadCell>
-        <TableHeadCell></TableHeadCell>
-        <TableHeadCell
-          >{$_("settings_versions_table_header_version")}</TableHeadCell
+<Table>
+  <TableHead class="bg-slate-400">
+    <TableHeadCell></TableHeadCell>
+    <TableHeadCell></TableHeadCell>
+    <TableHeadCell>{$_("settings_versions_table_header_version")}</TableHeadCell
+    >
+    <TableHeadCell>{$_("settings_versions_table_header_date")}</TableHeadCell>
+    <TableHeadCell>{$_("settings_versions_table_header_changes")}</TableHeadCell
+    >
+  </TableHead>
+  <TableBody class="divide-y *:text-white">
+    {#each releaseList as release (release.version)}
+      <TableBodyRow class="bg-slate-700">
+        <TableBodyCell class="px-6 py-2 whitespace-nowrap font-medium">
+          {#if release.isDownloaded}
+            <Radio
+              bind:group={versionState.activeToolingVersion}
+              value={release.version}
+              onchange={() => {
+                onVersionChange(release.version);
+              }}
+              disabled={!release.isDownloaded}
+              color="orange"
+              class="disabled:cursor-not-allowed p-0"
+            />
+          {/if}
+        </TableBodyCell>
+        <TableBodyCell
+          class="px-6 py-2 whitespace-nowrap font-medium"
+          style="line-height: 0;"
         >
-        <TableHeadCell
-          >{$_("settings_versions_table_header_date")}</TableHeadCell
-        >
-        <TableHeadCell
-          >{$_("settings_versions_table_header_changes")}</TableHeadCell
-        >
-      </TableHead>
-      <TableBody class="divide-y *:text-white">
-        {#each releaseList as release (release.version)}
-          <TableBodyRow class="bg-slate-700">
-            <TableBodyCell class="px-6 py-2 whitespace-nowrap font-medium">
-              {#if release.isDownloaded}
-                <Radio
-                  bind:group={versionState.activeToolingVersion}
-                  value={release.version}
-                  onchange={() => {
-                    onVersionChange(release.version);
-                  }}
-                  disabled={!release.isDownloaded}
-                  color="orange"
-                  class="disabled:cursor-not-allowed p-0"
+          <Button
+            class="py-0 bg-transparent focus:ring-0 disabled:opacity-50"
+            disabled={release.pendingAction}
+            onclick={async () => handleAction(release)}
+          >
+            {#if release.isDownloaded}
+              <IconDeleteForever
+                class="text-xl"
+                color="red"
+                aria-label={$_("settings_versions_icon_removeVersion_altText")}
+              />
+            {:else if release.pendingAction}
+              <Spinner color="yellow" size="6" />
+            {:else}
+              <IconDownload
+                class="text-xl"
+                color="#00d500"
+                aria-label={$_(
+                  "settings_versions_icon_downloadVersion_altText",
+                )}
+              />
+            {/if}
+          </Button>
+          {#if release.isDownloaded}
+            <Button
+              class="py-0 bg-transparent focus:ring-0 disabled:opacity-50"
+              disabled={release.pendingAction}
+              onclick={() => handleRedownload(release)}
+            >
+              {#if release.pendingAction}
+                <Spinner color="yellow" size="6" />
+              {:else}
+                <IconRefresh
+                  class="text-xl"
+                  aria-label={$_(
+                    "settings_versions_icon_redownloadVersion_altText",
+                  )}
                 />
               {/if}
-            </TableBodyCell>
-            <TableBodyCell
-              class="px-6 py-2 whitespace-nowrap font-medium"
-              style="line-height: 0;"
-            >
-              <Button
-                class="py-0 bg-transparent focus:ring-0 disabled:opacity-50"
-                disabled={release.pendingAction}
-                onclick={async () => handleAction(release)}
-              >
-                {#if release.isDownloaded}
-                  <IconDeleteForever
-                    class="text-xl"
-                    color="red"
-                    aria-label={$_(
-                      "settings_versions_icon_removeVersion_altText",
-                    )}
-                  />
-                {:else if release.pendingAction}
-                  <Spinner color="yellow" size="6" />
-                {:else}
-                  <IconDownload
-                    class="text-xl"
-                    color="#00d500"
-                    aria-label={$_(
-                      "settings_versions_icon_downloadVersion_altText",
-                    )}
-                  />
-                {/if}
-              </Button>
-              {#if release.isDownloaded}
-                <Button
-                  class="py-0 bg-transparent focus:ring-0 disabled:opacity-50"
-                  disabled={release.pendingAction}
-                  onclick={() => handleRedownload(release)}
-                >
-                  {#if release.pendingAction}
-                    <Spinner color="yellow" size="6" />
-                  {:else}
-                    <IconRefresh
-                      class="text-xl"
-                      aria-label={$_(
-                        "settings_versions_icon_redownloadVersion_altText",
-                      )}
-                    />
-                  {/if}
-                </Button>
-              {/if}
-            </TableBodyCell>
-            <TableBodyCell class="px-6 py-2 whitespace-nowrap font-medium">
-              {release.version}
-            </TableBodyCell>
-            <TableBodyCell class="px-6 py-2 whitespace-nowrap font-medium">
-              {#if release.date}
-                {new Date(release.date).toLocaleDateString()}
-              {/if}
-            </TableBodyCell>
-            <TableBodyCell class="px-6 py-2 whitespace-nowrap font-medium">
-              {#if release.githubLink}
-                <a href={release.githubLink} target="_blank" rel="noreferrer">
-                  <IconGitHub
-                    class="text-xl"
-                    aria-label={$_(
-                      "settings_versions_icon_githubRelease_altText",
-                    )}
-                  />
-                </a>
-              {/if}
-            </TableBodyCell>
-          </TableBodyRow>
-        {/each}
-      </TableBody>
-    </Table>
-  {/if}
-{/if}
+            </Button>
+          {/if}
+        </TableBodyCell>
+        <TableBodyCell class="px-6 py-2 whitespace-nowrap font-medium">
+          {release.version}
+        </TableBodyCell>
+        <TableBodyCell class="px-6 py-2 whitespace-nowrap font-medium">
+          {#if release.date}
+            {new Date(release.date).toLocaleDateString()}
+          {/if}
+        </TableBodyCell>
+        <TableBodyCell class="px-6 py-2 whitespace-nowrap font-medium">
+          {#if release.githubLink}
+            <a href={release.githubLink} target="_blank" rel="noreferrer">
+              <IconGitHub
+                class="text-xl"
+                aria-label={$_("settings_versions_icon_githubRelease_altText")}
+              />
+            </a>
+          {/if}
+        </TableBodyCell>
+      </TableBodyRow>
+    {/each}
+  </TableBody>
+</Table>
